+++ src/main/java/es/ujaen/dae/indicenciasurbanas/excepciones/UsuarioYaRegistrado.java
@@ -1,0 +1,7 @@
+package es.ujaen.dae.indicenciasurbanas.excepciones;
+
+public class UsuarioYaRegistrado extends RuntimeException {
+    public UsuarioYaRegistrado() {
+
+    }
+}